Use tools deps for testing and lein for packaging
It seems packaging the project as a jar for library use is difficult
using tools deps. There is only one tool that does this, meyvn, which
insists on collecting data about usage to be able to sell enterprise
licenses to heavy users. Not a very attractive feature. It also depends
on maven for secrets handling which means using maven's security setup
with encrypted passwords in the maven config file and a master password
rather than a gpg-encrypted edn file like leiningen and boot.

In sum I think using leiningen to package the library is preferable. We
can use the tools deps plugin for leiningen to fetch dependencies listed
in the deps.edn file. Publishing the library to nexus.nsd.no should be
done locally from developer machines (to sign package with gpg), so
having both leiningen and tools deps installed should be no problem.

Signed-off-by: Snorre Magnus Davøen <snorre.davoen@nsd.no>