Improve logging from resolve-key
The resolve-key function now logs first resolve attempt as a debug
statement. For usual INFO-level logging this will effectively turn off
logging of the first key resolve attempt which covers most resolve
calls.

If the key is not found the clj-jwt library will log the error with
error level before throwing the exception. This might help with
debugging when encountering problems with token validation.

Signed-off-by: Snorre Magnus Davøen <snorre.davoen@nsd.no>